refactor(coordinator): split audit substrate into storage and read model
audit_substrate.py held four change-reasons under one ownership: schema
definition, schema migration, record persistence, and analytics. A new
analytical query and a new record migration have nothing to do with each
other, but landed in the same module, were reviewed together, and collided
in the scheduler as a single file claim.

Split into two owners with a one-way dependency:

- audit_storage.py    — connection opening/validation, schema, the migration
                        catalogue, and every record write.
- audit_read_model.py — SELECT queries and the derivations over them.

The named interface is audit_storage.AuditConnection: a connection storage
has opened, validated, and brought to the current schema. The read model
issues SELECT SQL against one directly and calls storage's _load_migrated
when it needs a decoded record, so migration stays storage-owned. Storage
never imports the read model.

The 23-function migration catalogue moved as one uninterrupted sequential
unit and was not divided — it is meant to be read in order.

audit_substrate.py becomes a re-export facade, including the private names
(_meta_set, _migrate_record, _load_migrated, _open_validated, ...) that CLI
modules and tests import, so all ~20 consumers keep working unchanged.

verdict_outcome_counts() is the worked example that a new analytical query
needs no schema or migration change.

Verified against the real 923-record audit index: all 18 existing query
surfaces return byte-identical results pre- and post-split, and
rebuild_from_runs over 213 run records spanning 13 record schema versions
(9 through 28) produces identical rows in every table.

Refs #2350
